DCCT: Disable test function for fixed range option
Disable the test function option, which can configure the
DCCT sensor electronics to inject a calibration current
in the sensor, when the DCCT fixed range mode is selected.

Since the DCCT application with fixed range is fixed by
hardware, the test function is also fixed by hardware to
avoid different measurement conditions. The PV must therefore
be fixed so that it is consistent with the device state.